docs: fix example in README.md (#286)
Copy-pasted some example code in the UCAN workshop. This fix made it work: the invocations are each individual arguments, not an array.

Wondering though if the array form would be more intuitive when batching multiple invocations - in the old docs it would be `connection.execute(invocations)`, in the working code it's `connection.execute(...invocations)`.